requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
pandas>=2.1.0
plotly>=5.18.0
gunicorn>=21.2.0
google-generativeai>=0.7.0
